hybrid-assistant/requirements.txt
pillow
opencv-contrib-python
ultralytics
torch
transformers
accelerate
uvicorn[standard]
requests
numpy